requirements.txt
Beaker==1.6.4
Jinja2==2.7.1
Mako==1.0.0
MarkupSafe==0.18
SQLAlchemy==0.9.8
WTForms==2.0.1
alembic==0.6.7
argparse==1.2.1
bottle==0.11.6
bottle-cork==0.10
bottle-pgsql==0.2
bottle-sqlalchemy==0.4.1
bson==0.3.3
configobj==5.0.2
distribute==0.6.34
gevent==1.0.1
greenlet==0.4.5
gunicorn==18.0
psycopg2==2.5.4
pycrypto==2.6.1
pymongo==2.6.3
pytz==2013.9
six==1.5.2
wsgiref==0.1.2
